+++ spec/integrations/instrumentation/error_callback_subscription.rb
@@ -22,11 +22,11 @@ end
 event = error_events.first
 
 assert_not_equal 0, error_events.size
-assert_equal true, event.is_a?(Dry::Events::Event)
+assert event.is_a?(Dry::Events::Event)
 assert_equal 'error.occurred', event.id
 assert_equal 'librdkafka.error', event[:type]
-assert_equal true, event[:error].is_a?(Rdkafka::RdkafkaError)
+assert event[:error].is_a?(Rdkafka::RdkafkaError)
 assert_equal false, event[:subscription_group_id].empty?
 # We remap the consumer group id to a consumer group, thus the raw name is just a part of the
 # full name
-assert_equal true, event[:consumer_group_id].include?(DataCollector.consumer_groups.first)
+assert event[:consumer_group_id].include?(DataCollector.consumer_groups.first)