+++ spec/integrations/pro/licensing/configure_with_invalid_token.rb
@@ -12,7 +12,7 @@ rescue Karafka::Errors::InvalidLicenseTokenError
   failed_as_expected = true
 end
 
-assert_equal true, failed_as_expected
+assert failed_as_expected
 assert_equal false, Karafka.pro?
 
 # Pro components should not be visible