+++ spec/integrations/pro/rails/active_job/invalid_configuration_attempt.rb
@@ -34,10 +34,10 @@ handle { Job.karafka_options(partitioner: ->(job) { job.job_id }) }
 handle { Job.karafka_options(partitioner: Partitioner.new) }
 handle { Job.karafka_options(partitioner: NotPartitioner.new) }
 
-assert_equal true, DataCollector.data[0][0]
+assert DataCollector.data[0][0]
 assert_equal false, DataCollector.data[0][1]
-assert_equal true, DataCollector.data[0][2]
-assert_equal true, DataCollector.data[0][3]
+assert DataCollector.data[0][2]
+assert DataCollector.data[0][3]
 assert_equal false, DataCollector.data[0][4]
 assert_equal false, DataCollector.data[0][5]
-assert_equal true, DataCollector.data[0][6]
+assert DataCollector.data[0][6]