+++ spec/integrations/rails/rails6/just-a-dependency/rails_setup.rb
@@ -25,4 +25,4 @@ rescue Karafka::Errors::MissingBootFileError
   disabled = false
 end
 
-assert_equal true, disabled
+assert disabled