+++ spec/integrations/seek/seek_in_a_loop.rb
@@ -34,10 +34,10 @@ end
 
 # While we return there may be prefetched data that is still being processed before karafka stops
 # so we cannot have a strict limitation here (async)
-assert_equal true, DataCollector.data[0].size >= 420
+assert DataCollector.data[0].size >= 420
 
 # The last message should be consumed at least 5 times (first + min 4 loops)
-assert_equal true, DataCollector.data[0].count { |val| val == '1' } >= 5
+assert DataCollector.data[0].count { |val| val == '1' } >= 5
 
 # First 20 messages should be consumed only once
 elements[0..19].each do |payload|
@@ -46,7 +46,7 @@ end
 
 # All other messages should be consumed at least 5 times
 elements[20..-1].each do |payload|
-  assert_equal true, DataCollector.data[0].count { |val| val == payload } >= 5
+  assert DataCollector.data[0].count { |val| val == payload } >= 5
 end
 
 # Order of messages needs to be maintained within a single loop